src/backend/requirements.txt
python-dotenv
nylas
groq
flask